+++ kommunicateui/src/main/java/com/applozic/mobicomkit/uiwidgets/kommunicate/KmPrefSettings.java
@@ -113,7 +113,5 @@ public class KmPrefSettings {
         }
         return this;
     }
-
-
 }