+++ kommunicateui/src/main/java/com/applozic/mobicomkit/uiwidgets/kommunicate/callbacks/KmClickHandler.java
@@ -1,0 +1,7 @@
+package com.applozic.mobicomkit.uiwidgets.kommunicate.callbacks;
+
+import android.view.View;
+
+public interface KmClickHandler<T> {
+    void onItemClicked(View view, T data);
+}