+++ kommunicateui/src/main/java/com/applozic/mobicomkit/uiwidgets/AlCustomizationSettings.java
@@ -136,15 +136,21 @@ public class AlCustomizationSettings extends JsonMarker {
     private boolean restrictMessageTypingWithBots = false;
     private Map<String, Boolean> hidePostCTA;
     private boolean oneTimeRating;
+    private boolean restartConversationButtonVisibility = true;
+    private boolean rateConversationMenuOption;
     private boolean javaScriptEnabled;
-    private boolean restartConversationButtonVisibility;
-    
+
+    public boolean isRateConversationMenuOption() {
+        return rateConversationMenuOption;
+    }
+        
     public boolean isRestartConversationButtonVisibility() {
         return restartConversationButtonVisibility;
     }
     
     public boolean isJavaScriptEnabled() {
         return javaScriptEnabled;
+    }
 
     public Boolean getInnerTimestampDesign() {
         return innerTimestampDesign;