+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The changelog for [Kommunicate-Android-Chat-SDK](https://github.com/Kommunicate-
 2) Removed SQL injection vulnerability from the Android SDK
 3) Added support to set multiple Speech To Text language
 4) Added option to enable/disable JavaScript in WebView
+5) Restricted Zoom In / out on web view of FaqActivity
 
 
 ## Kommunicate Android SDK 2.6.2